+++ src/validation-verification-script/main.js
@@ -1,4 +1,4 @@
 goog.module('validationVerificationClientMain');
 const VerificationClient = goog.require('omid.verificationClient.VerificationClient');
 const ValidationVerificationClient = goog.require('omid.validationVerificationScript.ValidationVerificationClient');
-new ValidationVerificationClient(new VerificationClient(), 'dummyVendor');
+new ValidationVerificationClient(new VerificationClient(), 'iabtechlab.com-omid');